[FIX] FIxed the WorldToScreenPoint and WorldToViewportPoint  functions
