+++ Assets/Hotspot/Editor/LOD/CameraInfo.cs.meta
